move helper functions from test/mjolnir to test/util
