+++ test/mjolnir/test_dihedral_angle_interaction.cpp
@@ -1,7 +1,7 @@
 #define BOOST_TEST_MODULE "test_dihedral_angle_interaction"
 
 #include <boost/test/included/unit_test.hpp>
-#include <test/mjolnir/traits.hpp>
+#include <test/util/traits.hpp>
 #include <mjolnir/interaction/DihedralAngleInteraction.hpp>
 #include <mjolnir/math/constants.hpp>
 #include <mjolnir/potential/local/HarmonicPotential.hpp>
